+++ changelog.md
@@ -2,6 +2,17 @@
 
 > All Time Stamps: GMT+8
 
+## 0.2.42.0 _2020-02-03 10:00_
+
+* 增加方块和物品 added blocks and items
+  * 堇青石矿 cordierite ore block
+  * 堇青石 cordierite item
+  * 堇青石块 cordierite block
+* 增加材料 added materials
+  * 堇青石 cordierite
+* 增加属性 added traits
+  * 二色性 dichroic
+
 ## 0.2.41.0 _2020-02-01 14:30_
 
 * 更新语言文件 updated lang files